requirements.txt
google-cloud-videointelligence>=2.11.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
google-cloud-video-transcoder>=1.13.0
google-cloud-translate>=3.12.0
google-cloud-aiplatform>=1.60.0